remove old endMeeting
