+++ src/server/graphql/types/NotificationSubscriptionPayload.js
@@ -13,7 +13,6 @@ import AddFeatureFlagPayload from 'server/graphql/types/AddFeatureFlagPayload'
 import InviteToTeamPayload from 'server/graphql/types/InviteToTeamPayload'
 import AcceptTeamInvitationPayload from 'server/graphql/types/AcceptTeamInvitationPayload'
 import EndNewMeetingPayload from 'server/graphql/types/EndNewMeetingPayload'
-import EndMeetingPayload from 'server/graphql/types/EndMeetingPayload'
 import AddNewFeaturePayload from 'server/graphql/types/addNewFeaturePayload'
 
 const types = [
@@ -26,7 +25,6 @@ const types = [
   CreateTaskPayload,
   DeleteTaskPayload,
   DisconnectSocketPayload,
-  EndMeetingPayload,
   EndNewMeetingPayload,
   InviteToTeamPayload,
   RemoveOrgUserPayload,
